requirements.txt
discord.py
pynacl
pytube
python-dotenv
requests